fix(android): resolve WebSocket build issues
- Fix type mismatch: convert totalTime Long to Float
- Add packaging exclusions for Netty META-INF conflicts

+++ android/app/src/main/java/com/flam/edgeviewer/MainActivity.kt
@@ -314,7 +314,7 @@ class MainActivity : AppCompatActivity() {
                 height = height,
                 channels = channels,
                 fps = currentFps,
-                processingTimeMs = totalTime,
+                processingTimeMs = totalTime.toFloat(),
                 mode = modeString
             )
